chore: Update Google Places Autocomplete in HomeScreen
This commit updates the Google Places Autocomplete component in HomeScreen to include a predefined place for search. The 'Kyanja, Uganda' location is now included as a predefined place, allowing users to easily search for it. The previous import of the 'GOOGLE_MAPS_APIKEY' from '@env' has been commented out, as it is no longer needed.